stdnew

+++ ft_printf/src/print_di.c
@@ -19,7 +19,8 @@ char	*add_cut_di(char *snbr)
 
 	i = 0;
 	tmp = ft_strnew((ft_strlen(snbr) + (snbr[0] != '-') > (size_t)
-		g_flags->cut ? ft_strlen(snbr) + (snbr[0] != '-') : g_flags->cut) + 1);
+		g_flags->cut ? (int)ft_strlen(snbr) + (snbr[0] != '-') :
+		g_flags->cut) + 1);
 	if ((snbr)[0] != '-')
 	{
 		tmp[0] = '+';
